start replacing KF with nuke
